// src/enums/kit-weapon.ts
export enum KitWeapon {
	Bow = 'Лук',
	Ensnaring = 'Обплутувальна зброя',
	Heavy = 'Важка зброя',
	Light = 'Легка зброя',
	Medium = 'Середня зброя',
	Polearm = 'Древкова зброя',
	Unarmed = 'Беззбройний удар',
	Whip = 'Батіг'
}